+++ recorder_node/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog for package recorder_node
 
+tros_2.1.0 (2024-4-8)
+------------------
+1. 更新中英Readme说明。
+2. 适配humble版本rosbag2。
+
 tros_2.0.1 (2023-7-7)
+------------------
 1. 更新Readme说明。
 
 tros_2.0.0 (2023-6-8)